Allow overriding binaries installed by hack/go-install.sh
Makefile targets for binary dependencies (i.e. code-generator,
openshift-goimports and golangci-lint) now use version-less
filepaths:

    `$(TOOLS_DIR)/<Binary>` instead of `$(TOOLS_DIR)/<Binary>-<Version>`

`$(TOOLS_DIR)/<Binary>` may be a symlink to the versioned executable file.

Secondly, hack/go-install.sh now adds a check for `$(TOOLS_DIR)/<Binary>`
to be present and executable, in which case it skips `go install` invocation.

This allows for making local changes to these dependencies (e.g. when working
on kcp-dev/code-generator) and generating client code.

On-behalf-of: SAP robert.vasek@sap.com

+++ hack/go-install.sh
@@ -49,6 +49,9 @@ function clean {
 }
 trap clean EXIT
 
+# An executable already exists, exit early.
+[[ -f "${GOBIN}/${2}" && -x $(realpath "${GOBIN}/${2}") ]] && exit 0
+
 rm "${GOBIN}/${2}"* > /dev/null 2>&1 || true
 
 cd "${tmp_dir}"
